+++ docs/changelog.rst
@@ -8,6 +8,7 @@
 - TwitchIO
     - Loosen aiohttp requirements to allow 3.8.1
     - :class:`Stream` was missing from ``__all__``. It is now available in the twitchio namespace.
+    - Added ``.status`` and ``.reason`` to :class:`HTTPException`
   
 - ext.pubsub
     - Add channel subscription pubsub model